+++ packages/moonstone/RangePicker/RangePicker.js
@@ -1,6 +1,8 @@
 /**
- * Exports the {@link moonstone/RangePicker.RangePicker} and
- * {@link moonstone/RangePicker.RangePickerBase} components
+ * A component for selecting a number from a range of numbers.
+ *
+ * @example
+ * <RangePicker defaultValue={70} min={0} max={100}></RangePicker>
  *
  * @module moonstone/RangePicker
  */
@@ -24,9 +26,8 @@ const digits = (num) => {
 };
 
 /**
- * {@link moonstone/RangePicker.RangePickerBase} is a component that lets the user select a number
- * from a range of numbers. This version is not spottable. Developers are encouraged to use
- * {@link moonstone/RangePicker.RangePicker}.
+ * Base component of [RangePicker]{@link moonstone/RangePicker.RangePicker} which is not
+ * [spottable]{@link spotlight/Spottable.Spottable}.
  *
  * @class RangePickerBase
  * @memberof moonstone/RangePicker
@@ -38,7 +39,7 @@ const RangePickerBase = kind({
 
 	propTypes: /** @lends moonstone/RangePicker.RangePickerBase.prototype */ {
 		/**
-		 * The maximum value selectable by the picker (inclusive).
+		 * Maximum selectable value (inclusive).
 		 *
 		 * @type {Number}
 		 * @required
@@ -47,7 +48,7 @@ const RangePickerBase = kind({
 		max: PropTypes.number.isRequired,
 
 		/**
-		 * The minimum value selectable by the picker (inclusive).
+		 * Minimum selectable value (inclusive).
 		 *
 		 * @type {Number}
 		 * @required
@@ -56,7 +57,7 @@ const RangePickerBase = kind({
 		min: PropTypes.number.isRequired,
 
 		/**
-		 * Current value
+		 * Current value.
 		 *
 		 * @type {Number}
 		 * @required
@@ -65,8 +66,9 @@ const RangePickerBase = kind({
 		value: PropTypes.number.isRequired,
 
 		/**
-		 * Overrides the `aria-valuetext` for the picker. By default, `aria-valuetext` is set
-		 * to the current selected child value.
+		 * The `aria-valuetext` for the picker.
+		 *
+		 * By default, `aria-valuetext` is set to the current selected child value.
 		 *
 		 * @type {String}
 		 * @memberof moonstone/RangePicker.RangePickerBase.prototype
@@ -75,7 +77,7 @@ const RangePickerBase = kind({
 		'aria-valuetext': PropTypes.string,
 
 		/**
-		 * Children from which to pick
+		 * Children from which to pick.
 		 *
 		 * @type {Node}
 		 * @public
@@ -83,7 +85,7 @@ const RangePickerBase = kind({
 		children: PropTypes.node,
 
 		/**
-		 * Class name for component
+		 * Class name for component.
 		 *
 		 * @type {String}
 		 * @public
@@ -91,9 +93,11 @@ const RangePickerBase = kind({
 		className: PropTypes.string,
 
 		/**
-		 * Assign a custom icon for the decrementer. All strings supported by [Icon]{@link moonstone/Icon.Icon} are
-		 * supported. Without a custom icon, the default is used, and is automatically changed when
-		 * the [orientation]{@link moonstone/Icon.Icon#orientation} is changed.
+		 * A custom icon for the decrementer.
+		 *
+		 * All strings supported by [Icon]{@link moonstone/Icon.Icon} are supported. Without a
+		 * custom icon, the default is used, and is automatically changed when the
+		 * [orientation]{@link moonstone/RangePicker.RangePicker#orientation} is changed.
 		 *
 		 * @type {string}
 		 * @public
@@ -101,8 +105,7 @@ const RangePickerBase = kind({
 		decrementIcon: PropTypes.string,
 
 		/**
-		 * When `true`, the RangePicker is shown as disabled and does not generate `onChange`
-		 * [events]{@glossary event}.
+		 * Disables the picker.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -110,9 +113,11 @@ const RangePickerBase = kind({
 		disabled: PropTypes.bool,
 
 		/**
-		 * Assign a custom icon for the incrementer. All strings supported by [Icon]{@link moonstone/Icon.Icon} are
-		 * supported. Without a custom icon, the default is used, and is automatically changed when
-		 * the [orientation]{@link moonstone/Icon.Icon#orientation} is changed.
+		 * A custom icon for the incrementer.
+		 *
+		 * All strings supported by [Icon]{@link moonstone/Icon.Icon} are supported. Without a
+		 * custom icon, the default is used, and is automatically changed when the
+		 * [orientation]{@link moonstone/RangePicker.RangePicker#orientation} is changed.
 		 *
 		 * @type {String}
 		 * @public
@@ -120,11 +125,11 @@ const RangePickerBase = kind({
 		incrementIcon: PropTypes.string,
 
 		/**
-		 * Determines the user interaction of the control. A joined picker allows the user to use
-		 * the arrow keys to adjust the picker's value. The user may no longer use those arrow keys
-		 * to navigate, while this control is focused. A split control allows full navigation,
-		 * but requires individual ENTER presses on the incrementer and decrementer buttons.
-		 * Pointer interaction is the same for both formats.
+		 * Allows the user can use the arrow keys to adjust the picker's value.
+		 *
+		 * The user may no longer use those arrow keys to navigate while this control is focused.
+		 * A default control allows full navigation, but requires individual ENTER presses on the
+		 * incrementer and decrementer buttons. Pointer interaction is the same for both formats.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -132,9 +137,10 @@ const RangePickerBase = kind({
 		joined: PropTypes.bool,
 
 		/**
-		 * By default, the picker will animate transitions between items if it has a defined
-		 * `width`. Specifying `noAnimation` will prevent any transition animation for the
-		 * component.
+		 * Disables animation.
+		 *
+		 * By default, the picker will animate transitions between items, provided a `width` is
+		 * defined.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -142,7 +148,7 @@ const RangePickerBase = kind({
 		noAnimation: PropTypes.bool,
 
 		/**
-		 * A function to run when the control should increment or decrement.
+		 * Called when `value` changes.
 		 *
 		 * @type {Function}
 		 * @public
@@ -150,8 +156,11 @@ const RangePickerBase = kind({
 		onChange: PropTypes.func,
 
 		/**
-		 * Sets the orientation of the picker, whether the buttons are above and below or on the
-		 * sides of the value. Must be either `'horizontal'` or `'vertical'`.
+		 * Orientation of the picker.
+		 *
+		 * Controls whether the buttons are arranged horizontally or vertically around the value.
+		 *
+		 * * Values: `'horizontal'`, `'vertical'`
 		 *
 		 * @type {String}
 		 * @default 'horizontal'
@@ -160,8 +169,8 @@ const RangePickerBase = kind({
 		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
 
 		/**
-		 * When `true`, pads the display value with zeros up to the number of digits of the value of
-		 * `min` or max`, whichever is greater.
+		 * Pads the display value with zeros up to the number of digits of `min` or max`, whichever
+		 * is greater.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -169,8 +178,9 @@ const RangePickerBase = kind({
 		padded: PropTypes.bool,
 
 		/**
-		 * Allow the picker to only increment or decrement by a given value. A step of `2` would
-		 * cause a picker to increment from 10 to 12 to 14, etc.
+		 * The smallest value change allowed for the picker.
+		 *
+		 * For example, a step of `2` would cause the picker to increment from 0 to 2 to 4, etc.
 		 *
 		 * @type {Number}
 		 * @default 1
@@ -179,13 +189,19 @@ const RangePickerBase = kind({
 		step: PropTypes.number,
 
 		/**
-		 * Choose a specific size for your picker. `'small'`, `'medium'`, `'large'`, or set to `null` to
-		 * assume auto-sizing. `'small'` is good for numeric pickers, `'medium'` for single or short
-		 * word pickers, `'large'` for maximum-sized pickers.
+		 * The width of the picker.
 		 *
-		 * You may also supply a number. This number will determine the minumum size of the Picker.
-		 * Setting a number to less than the number of characters in your longest value may produce
-		 * unexpected results.
+		 * A number can be used to set the minimum number of characters to be shown. Setting a
+		 * number to less than the number of characters in the longest value will cause the width to
+		 * grow for the longer values.
+		 *
+		 * A string can be used to select from pre-defined widths:
+		 *
+		 * * `'small'` - numeric values
+		 * * `'medium'` - single or short words
+		 * * `'large'` - maximum-sized pickers taking full width of its parent
+		 *
+		 * By default, the picker will size according to the longest valid value.
 		 *
 		 * @type {String|Number}
 		 * @public
@@ -196,8 +212,8 @@ const RangePickerBase = kind({
 		]),
 
 		/**
-		 * Should the picker stop incrementing when the picker reaches the last element? Set `wrap`
-		 * to true to allow the picker to continue from the opposite end of the list of options.
+		 * Allows picker to continue from the start of the list after it reaches the end and
+		 * vice-versa.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -225,13 +241,16 @@ const RangePickerBase = kind({
 				validateRange(value, min, max, 'RangePicker');
 			}
 			return clamp(min, max, value);
+		},
+		voiceLabel: ({min, max}) => {
+			return JSON.stringify([min, max]);
 		}
 	},
 
-	render: ({label, value, ...rest}) => {
+	render: ({label, value, voiceLabel, ...rest}) => {
 		delete rest.padded;
 		return (
-			<Picker {...rest} index={0} value={value} reverse={false}>
+			<Picker {...rest} data-webos-voice-labels-ext={voiceLabel} index={0} value={value} reverse={false}>
 				<PickerItem key={value} marqueeDisabled style={{direction: 'ltr'}}>{label}</PickerItem>
 			</Picker>
 		);
@@ -239,8 +258,7 @@ const RangePickerBase = kind({
 });
 
 /**
- * {@link moonstone/RangePicker.RangePicker} is a component that lets the user select a number from
- * a range of numbers.
+ * A component that lets the user select a number from a range of numbers.
  *
  * By default, `RangePicker` maintains the state of its `value` property. Supply the `defaultValue`
  * property to control its initial value. If you wish to directly control updates to the component,
@@ -257,6 +275,15 @@ const RangePicker = Pure(
 		RangePickerBase
 	)
 );
+
+/**
+ * Default value
+ *
+ * @name defaultValue
+ * @memberof moonstone/RangePicker.RangePicker.prototype
+ * @type {Number}
+ * @public
+ */
 
 export default RangePicker;
 export {RangePicker, RangePickerBase};