+++ packages/sampler/stories/moonstone-stories/Slider.js
@@ -7,7 +7,8 @@ import {withInfo} from '@storybook/addon-info';
 import {boolean, number, select} from '../../src/enact-knobs';
 import {mergeComponentMetadata} from '../../src/utils';
 
-const Config = mergeComponentMetadata('Slider', Slider);
+const SliderConfig = mergeComponentMetadata('Slider', Slider);
+const SliderTooltipConfig = mergeComponentMetadata('SliderTooltip', SliderTooltip);
 
 Slider.displayName = 'Slider';
 SliderTooltip.displayName = 'SliderTooltip';
@@ -19,22 +20,22 @@ storiesOf('Moonstone', module)
 			propTablesExclude: [Slider, SliderTooltip],
 			text: 'Basic usage of Slider'
 		})(() => {
-			const side = select('side', ['after', 'before', 'left', 'right'], SliderTooltip, 'before');
-			const tooltip = boolean('tooltip', SliderTooltip);
-			const percent = boolean('percent', SliderTooltip);
+			const side = select('side', ['after', 'before', 'left', 'right'], SliderTooltipConfig, 'before');
+			const tooltip = boolean('tooltip', SliderTooltipConfig);
+			const percent = boolean('percent', SliderTooltipConfig);
 
 			return (
 				<Slider
-					activateOnFocus={boolean('activateOnFocus', Config)}
-					backgroundProgress={number('backgroundProgress', Config, {range: true, min: 0, max: 1, step: 0.01}, 0.5)}
-					disabled={boolean('disabled', Config)}
-					knobStep={number('knobStep', Config)}
-					max={number('max', Config, 10)}
-					min={number('min', Config, 0)}
-					noFill={boolean('noFill', Config)}
+					activateOnFocus={boolean('activateOnFocus', SliderConfig)}
+					backgroundProgress={number('backgroundProgress', SliderConfig, {range: true, min: 0, max: 1, step: 0.01}, 0.5)}
+					disabled={boolean('disabled', SliderConfig)}
+					knobStep={number('knobStep', SliderConfig)}
+					max={number('max', SliderConfig, 10)}
+					min={number('min', SliderConfig, 0)}
+					noFill={boolean('noFill', SliderConfig)}
 					onChange={action('onChange')}
-					orientation={select('orientation', ['horizontal', 'vertical'], Config, 'horizontal')}
-					step={number('step', Config, 1)}
+					orientation={select('orientation', ['horizontal', 'vertical'], SliderConfig, 'horizontal')}
+					step={number('step', SliderConfig, 1)}
 				>
 					{tooltip ? (
 						<SliderTooltip