+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to track pointer mode while paused
+
+## [2.0.0-rc.2] - 2018-07-16
+
+### Added
+
+- `spotlight` debugging to visualize which components will be targeted as the next component for any 5-way direction
+
+## [2.0.0-rc.1] - 2018-07-09
+
+No significant changes.
+
 ## [2.0.0-beta.9] - 2018-07-02
 
 No significant changes.